requirements.txt
langchain
python-dotenv
langchain_openai
unstructured
langchain_community